+++ pkg/infra/cmdLine/cmdLineInits.go
@@ -4,6 +4,7 @@ type CmdLine struct {
 	FixedSanctionDuration int
 	DynamicSanctions      bool
 	GraduatedSanctions    bool
+	PersistentSanctions   bool
 }
 
 var CmdLineInits CmdLine = CmdLine{}
